--- .nuget/nuget.config
@@ -1,7 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <packageSources>
-    <add key="nuget.org" value="https://api.nuget.org/v3/index.json" />
-    <add key="Episerver" value="https://nuget.episerver.com/feed/packages.svc/" />
-  </packageSources>
-</configuration>
