--- src/DoubleJay.Epi.EnhancedPropertyList/DoubleJay.Epi.EnhancedPropertyList.nuspec
@@ -1,23 +1,0 @@
-<?xml version="1.0"?>
-<package>
-  <metadata>
-    <id>$id$</id>
-    <version>$version$</version>
-    <title>$title$</title>
-    <authors>$author$</authors>
-    <owners>$author$</owners>
-    <projectUrl>https://github.com/jacobjones/DoubleJay.Epi.EnhancedPropertyList</projectUrl>
-    <licenseUrl>https://github.com/jacobjones/DoubleJay.Epi.EnhancedPropertyList/blob/master/LICENSE</licenseUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>$description$</description>
-    <releaseNotes></releaseNotes>
-    <iconUrl>https://raw.githubusercontent.com/jacobjones/DoubleJay.Epi.EnhancedPropertyList/master/icons/enhanced-property-list.png</iconUrl>
-    <dependencies>
-      <dependency id="EPiServer.CMS.Core" version="[11.1.0,12.0.0)" />
-      <dependency id="EPiServer.CMS.UI.Core" version="[11.1.0,12.0.0)" />
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="bin\DoubleJay.Epi.EnhancedPropertyList.zip" target="content\modules\_protected\DoubleJay.Epi.EnhancedPropertyList" />
-  </files>
-</package>
