--- src/DoubleJay.Epi.EnhancedPropertyList/Properties/AssemblyInfo.cs
@@ -1,35 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("DoubleJay.Epi.EnhancedPropertyList")]
-[assembly: AssemblyDescription("Enhances the generic PropertyList in Episerver to visually improve Urls, ContentReferences and images")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Jake Jones")]
-[assembly: AssemblyProduct("DoubleJay.Epi.EnhancedPropertyList")]
-[assembly: AssemblyCopyright("Copyright © Jake Jones 2020")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-
-// Setting ComVisible to false makes the types in this assembly not visible
-// to COM components.  If you need to access a type in this assembly from
-// COM, set the ComVisible attribute to true on that type.
-[assembly: ComVisible(false)]
-
-// The following GUID is for the ID of the typelib if this project is exposed to COM
-[assembly: Guid("023a31a6-4636-4066-8af2-f18f62bbd30c")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.1")]
-[assembly: AssemblyFileVersion("1.0.1")]
