--- src/DoubleJay.Epi.EnhancedPropertyList/module.config
@@ -1,16 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<module loadFromBin="false" name="DoubleJay.Epi.EnhancedPropertyList" clientResourceRelativePath="$version$" tags="EPiServerModulePackage">
-  <assemblies>
-    <add assembly="DoubleJay.Epi.EnhancedPropertyList" />
-  </assemblies>
-  <clientModule>
-    <moduleDependencies>
-      <add dependency="CMS" type="RunAfter" />
-    </moduleDependencies>
-  </clientModule>
-  <dojo>
-    <paths>
-      <add name="enhancedpropertylist" path="ClientResources/Scripts" />
-    </paths>
-  </dojo>
-</module>
