+++ src/DoubleJay.Epi.EnhancedPropertyList/module/module.config
@@ -1,0 +1,16 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<module loadFromBin="false" name="DoubleJay.Epi.EnhancedPropertyList" clientResourceRelativePath="">
+  <assemblies>
+    <add assembly="DoubleJay.Epi.EnhancedPropertyList" />
+  </assemblies>
+  <clientModule>
+    <moduleDependencies>
+      <add dependency="CMS" type="RunAfter" />
+    </moduleDependencies>
+  </clientModule>
+  <dojo>
+    <paths>
+      <add name="enhancedpropertylist" path="ClientResources/Scripts" />
+    </paths>
+  </dojo>
+</module>
