--- src/DoubleJay.Epi.EnhancedPropertyList/packages.config
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Castle.Core" version="4.2.1" targetFramework="net461" developmentDependency="true" />
-  <package id="Castle.Windsor" version="4.1.0" targetFramework="net461" />
-  <package id="EPiServer.CMS.AspNet" version="11.1.0" targetFramework="net461" />
-  <package id="EPiServer.CMS.Core" version="11.1.0" targetFramework="net461" />
-  <package id="EPiServer.CMS.UI.Core" version="11.1.0" targetFramework="net461" />
-  <package id="EPiServer.Framework" version="11.1.0" targetFramework="net461" />
-  <package id="EPiServer.Framework.AspNet" version="11.1.0" targetFramework="net461" />
-  <package id="Microsoft.AspNet.Mvc" version="4.0.20710.0" targetFramework="net461" />
-  <package id="Microsoft.AspNet.Razor" version="2.0.20710.0" targetFramework="net461" />
-  <package id="Microsoft.AspNet.WebPages" version="2.0.20710.0" targetFramework="net461" />
-  <package id="Microsoft.Tpl.Dataflow" version="4.5.24" targetFramework="net461" />
-  <package id="Microsoft.Web.Infrastructure" version="1.0.0.0" targetFramework="net461" />
-  <package id="Newtonsoft.Json" version="9.0.1" targetFramework="net461" />
-  <package id="structuremap.web-signed" version="3.1.6.186" targetFramework="net461" developmentDependency="true" />
-  <package id="structuremap-signed" version="3.1.6.186" targetFramework="net461" developmentDependency="true" />
-  <package id="System.ComponentModel.Annotations" version="4.4.0" targetFramework="net461" />
-  <package id="System.Data.SqlClient" version="4.4.0" targetFramework="net461" />
-  <package id="System.Reflection.Emit" version="4.3.0" targetFramework="net461" />
-  <package id="System.Security.AccessControl" version="4.4.0" targetFramework="net461" />
-  <package id="System.Security.Cryptography.Xml" version="4.4.0" targetFramework="net461" />
-  <package id="System.Security.Permissions" version="4.4.0" targetFramework="net461" />
-  <package id="System.Security.Principal.Windows" version="4.4.0" targetFramework="net461" />
-  <package id="System.Threading.AccessControl" version="4.4.0" targetFramework="net461" />
-</packages>
